More readable code for File class
Instead of dynamically figuring out how to calculate fields, this
creates a set of new properties that each one take care of a single job.
This not only helps with code readability, but also helps on
encapsulating the logic better.

In most places I tried to keep the logic as it was, so all tests should
still pass. Still, some changes are included that take advantage of some
Python good practices

- Use context managers to open files, in that way, any errors while
  reading the files or updating the hash objects will automatically
  close the file objects.
- Instead of setting properties on the fly, this defines properties as
  custom methods that take care of calculating values or caching them
  internally in the object